requirements.txt
# Core
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.9.0
pydantic-settings>=2.5.0
email-validator>=2.2.0

# Database
sqlalchemy[asyncio]>=2.0.35
asyncpg>=0.30.0
alembic>=1.14.0
pgvector>=0.3.5

# Redis
redis>=5.2.0

# HTTP / Crawling
aiohttp>=3.11.0
aiohttp-socks>=0.9.0
# OTX, AbuseIPDB, and several Cloudflare/Akamai-fronted upstreams
# negotiate ``brotli`` or ``zstandard`` encoding regardless of our
# Accept-Encoding hint. Without these decoders aiohttp surfaces the
# response as a ``ClientResponseError 400 (Can not decode content-
# encoding)`` even when the upstream sent a valid 200.
brotli>=1.1.0
zstandard>=0.23.0
beautifulsoup4>=4.12.0
# playwright>=1.49.0  # Only needed for crawler mode, not API-only deploys

# LLM (all optional — only install what you need)
# ollama is accessed via HTTP, no package needed

# PDF Reports
reportlab>=4.1.0
# Compliance Evidence Pack (P1 #1.3) — Arabic shaping + bidi reordering
# for bilingual exports. Both pure-Python, no native deps.
arabic-reshaper>=3.0.0
python-bidi>=0.6.0

# Sigma rule generation + multi-SIEM translation (P2 #2.3). pysigma
# pure-Python; backend modules pinned to stable releases. Each backend
# emits one of Splunk SPL / Elastic Lucene+ES|QL / KQL (Sentinel,
# M365Defender) / IBM QRadar AQL.
# pySigma — pinned to exact versions to avoid pip's resolver exploring
# every minor combination across the six packages. The combination of
# pysigma + 5 backends + stix-shifter + 4 modules + pycti expanded to
# the resolver's max_rounds (200000) when left unbounded. Pin via the
# active set in the host venv (./.venv/bin/pip freeze).
pysigma==0.11.23
pysigma-backend-splunk==1.1.3
pysigma-backend-elasticsearch==1.1.6
pysigma-backend-kusto==0.4.5
pysigma-backend-microsoft365defender==0.3.2
pysigma-backend-qradar-aql==0.3.2

# STIX-Shifter (P2 #2.5) — translate STIX 2.x patterns into native query
# language for ~35 data sources. Pinned to 8.1.1 — modules MUST match
# the main package version; with ``>=8.1.0`` pip explored every minor
# release combination.
stix-shifter==8.1.1
stix-shifter-utils==8.1.1
stix-shifter-modules-splunk==8.1.1
stix-shifter-modules-elastic-ecs==8.1.1
stix-shifter-modules-qradar==8.1.1
stix-shifter-modules-azure-sentinel==8.1.1

# YARA-X + Mandiant capa (P2 #2.10). yara-x is the Rust-rewrite of YARA
# with a drop-in Python binding (~5x faster compile + scan vs yara-python);
# capa extracts adversary capabilities from PE / ELF / Mach-O binaries.
yara-x==1.15.0
flare-capa==9.4.0

# PyMISP — Argus calls customer-private + CIRCL public MISP servers
# (P2 #2.11). We consume the API only; the AGPL MISP server itself is
# not bundled.
pymisp==2.5.34.1

# pycti — read/write to a co-deployed OpenCTI instance (P2 #2.1).
# Apache-2.0; we don't bundle the OpenCTI server, the operator runs
# their own. The wrapper detects unavailability and degrades to no-op.
# pycti's release cadence is daily (date-stamped versions like
# 7.260430.0); pin to a known-good snapshot.
pycti==7.260430.0

# Notifications
aiosmtplib>=3.0.0

# Multipart form parsing (for /assets/bulk/csv upload)
python-multipart>=0.0.20

# Evidence Vault — MinIO/S3-compatible blob storage
boto3>=1.35.0

# Notifications — webhooks, SMTP, and Jasmin SMS gateway
aiohttp>=3.11.0
aiosmtpd>=1.4.6

# EASM — DNS + WHOIS for runner pipelines
dnspython>=2.7.0
python-whois>=0.9.5

# Hardening — XXE-safe XML parsing + magic-byte file sniffing
defusedxml>=0.7.1
python-magic>=0.4.27

# DMARC360 — RUA aggregate report parsing. parsedmarc 9.x pulls in
# many transitive deps; pin so pip's resolver doesn't have to explore.
parsedmarc==9.11.0

# Brand Protection — typosquat / lookalike domain generator
dnstwist==20250130

# OSINT / leakage — does this email exist on 120+ services?
# Powers /leakage exposure surface alongside HudsonRock Cavalier.
holehe>=1.61

# Apprise — OSS-default notification fan-out. Single library, one URL
# scheme per service (slack://, mailto://, ntfy://, mattermost://,
# rocket://, pagerduty://, telegram://, discord://, etc.) — operators
# don't need to wire each notifier individually.
apprise>=1.9.0

# theHarvester is published on PyPI under a different cadence and
# the latest stable version isn't reliably mirrored. Install at boot
# via ``pip install git+https://github.com/laramies/theHarvester``
# from the Dockerfile post-pin (see Dockerfile + RUN steps) instead
# of pinning here.

# Brand Protection — TLD parsing driven by the Public Suffix List.
# tldextract ships a bundled PSL snapshot; ``include_psl_private_domains``
# keeps the snapshot deterministic in offline / air-gap deploys.
tldextract>=5.1.2

# Logo abuse — perceptual hashing (lightweight CLIP alternative — see docs/HARDWARE_DECISIONS.md)
imagehash>=4.3.2
Pillow>=11.0.0

# Evidence Vault — text extraction for the artefact summariser agent.
# pypdf parses PDFs; pytesseract is a thin wrapper over the optional
# Tesseract binary (degrades gracefully when the binary is absent).
pypdf>=5.0.0
pytesseract>=0.3.13

# Impersonation matching — fast string + name similarity
rapidfuzz>=3.10.0

# Auth
argon2-cffi>=25.1.0
PyJWT>=2.9.0
cryptography>=43.0.0

# Geo / Feeds
geoip2>=4.8.0
websockets>=12.0

# Local security tools
yara-python>=4.5.0
PyYAML>=6.0.0

# Utils
python-dotenv>=1.0.0

# Observability (Audit C6) — pure-Python, ~64 KB, no native deps
prometheus_client>=0.20.0

# 2FA / TOTP (Audit D10) — pure-Python
pyotp>=2.9.0

# Mobile-app store scrapers (Audit B3 — Phase 4.4)
#   Android: native aiohttp scraper in src/social/mobile_apps.py
#       — replaced the third-party google-play-scraper lib because it
#       does brittle deep-DOM traversal that returns None for short
#       queries (e.g. "ENBD"), silently dropping coverage. We now
#       parse Open Graph meta tags from /store/apps/details pages,
#       which is a stable web standard.
#   iOS: Apple's public iTunes Search API hit directly with aiohttp.

# Social monitoring (Audit B3 — Phase 4.1) — open-source scrapers
# that work without paid API keys. Each is opt-in per worker config.
# Pinned because each one has loose transitive specs that fight the
# rest of the dependency graph during pip resolution.
#   Instagram: instaloader (Bellingcat-recommended OSINT tool, public
#              profiles work without login at moderate rate limits).
#   TikTok: davidteather/TikTok-Api — actively maintained, used by
#          250+ companies, MIT license. Requires Playwright Chromium
#          to be installed (`playwright install chromium`). Opt-in
#          via ARGUS_WORKER_TIKTOK_INTERVAL > 0; default 0 (off).
instaloader==4.15.1
TikTokApi==7.3.3
#   Twitter/X: Scweet — replays GraphQL via curl_cffi w/ TLS
#              fingerprinting; multi-account pooling, proxy support.
#              Operator must seed authenticated sessions via Scweet's
#              account-session mechanism. Opt-in via
#              ARGUS_WORKER_TWITTER_INTERVAL > 0 + populated session dir.
Scweet==5.3
#   LinkedIn: linkedin-scraper — Selenium-based, requires
#             ARGUS_LINKEDIN_USERNAME/PASSWORD operator credentials,
#             very aggressive ban risk: limit to 1 req / 60s and only
#             scan Company pages (legally scrapable per hiQ Labs v.
#             LinkedIn 9th Circuit ruling). Opt-in via
#             ARGUS_WORKER_LINKEDIN_INTERVAL > 0.
linkedin-scraper>=3.1.0

# Test harness (Phase 0 — required for every later phase)
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0
httpx>=0.27.0
asgi-lifespan>=2.1.0
